requirements.txt
scrapy
psycopg2-binary
elasticsearch==8.15.0
w3lib
python-dotenv